requirements.txt
numpy==1.24.3
matplotlib==3.7.1
scikit-learn==1.2.2
tqdm==4.65.0
google-auth-oauthlib
google-auth
google-api-python-client 